backend/requirements.txt
flask
flask-cors
google-generativeai
scikit-learn
torch
transformers
pandas
numpy
joblib
twilio
requests
langchain
langchain-google-genai
langchain-anthropic
langchain-core
langchain-openai
langchain-cohere
langchain-huggingface
python-dotenv
gunicorn
